networkmanager: use modify_system=true
This does the following:
Allow users to modify system connections [0].

Techinically what it does is change the NM_MODIFY_SYSTEM_POLICY
to "yes" instead of "auth_admin_keep" which would require a password prompt.
That is inacted in the following polkit policy [1].

This allows us to drop a permissive looking polkit rule, and we don't
have to instruct users to add themselves to the "networkmanager" group.

All relevant documentation is updated.

[0]: https://gitlab.freedesktop.org/NetworkManager/NetworkManager/blob/1.20.4/meson_options.txt#L17
[1]: https://gitlab.freedesktop.org/NetworkManager/NetworkManager/blob/1.20.4/data/org.freedesktop.NetworkManager.policy.in.in#L115