automake-1.15: deterministic documentation generation
automake 1.15 uses a bundled version of help2man from 2012 that
does not support the SOURCE_DATE_EPOCH environment variable.

This means that the build date is included in the generated
documentation, breaking reproducibility.

This changes adds the SOURCE_DATE_EPOCH support from the current
help2man to the version bundled in automake 1.15.