emacsPackages.vterm: fix dynamic library and autoloaders
We need both the .so as well as the proper autoloaders that are generated when
we build an emacs package, so we cannot use the normal stdenv.mkDerivation but
have to inject the building of the libraries instead.

Also, use the proper libvterm-neovim we have in nixpkgs instead of vendoring it.