execlineb: change execlineb wrapper to C script
Instead of using execlineb to define the execlineb wrapper, we replace
it by a little C wrapper.

This is mainly done because on non-Linux systems (i.e. mainly macOS),
it is impossible for a shebang interpreter to be itself a shebang
script.
It is, however, perfectly fine to have a chain that goes
shebang -> ELF -> shebang -> ELF -> …

Co-Authored-By: Laurent Bercot <ska-skaware@skarnet.org>