vcv-rack: Fix issues from missing upstream dependency
The revision of the glfw fork that Rack depends on has been removed from
GitHub. This causes cloning of the glfw submodule for the Rack git
repository to fail.

As the glfw fork just adds a function that is only needed on Darwin (and
we don't support that platform at the moment) we remove the function
call from the Rack source. We can then use the upstream version of glfw.

We fetch the rest of the submodules manually as well.

See https://github.com/NixOS/nixpkgs/issues/71189 for the context.