release-lib.nix: initialize pkgs for the currentSystem
Previously calling release-lib for only x86_64-darwin like:

  import ./release-lib.nix {
    supportedSystems = [ "x86_64-darwin" ];
  }

would still cause the packageSet to be evaluated for
x86_64-linux. This is annoying if you are evaluating this on
x86_64-darwin and your packageSet uses Import From Derivation to
produce a jobset. In that case you need to have a x86_64-linux build
machine configured in order to finish the evaluation.

This commit initializes `pkgs` for the currentSystem such that all IFD
derivations are build for the current system.