python.pkgs.twisted: move twisted cache generation to postFixup
The Python 2 build of twisted would fail when generating the cache. This
failure started occuring since the change was introduced that builds
setuptools as a wheel instead of as an egg,
90be4c2c7875c9487508d95b5c638d97e2903ada.

Turns out moving this step to postFixup solves the issue. I suppose it
is related to the order of site packages being evaluated, where
PYTHONPATH introduced additional ones.