qtkeychain: correctly define buildInputs
Because of a small mistake in the use of lib.lists.optional [^1],
libsecret was not being used during the build of qtkeychain. This commit
fixes that mistake. But since including libsecret apparently breaks some
packages, this commit makes the inclusion of libsecret optional via the
withLibsecret option, which defaults to false. Therefore, this change
should be free of any consequences.

Note that this commit also changes one occurrence of lib.lists.optional
that is used for Darwin-specific buildInputs. I have tested (with
nix-review) whether this breaks anything on macOS. Everything seems to
be fine. The output of nix-review has been: “Nothing changed."

[^1]: lib.lists.optional wraps its input in a list. Therefore, the argument to
this function should usually not be a list because then the result is a
nested list, which is, in most cases, not the desired output.
lib.lists.optionals — note the ’s’ at the end — on the other hand does
not wrap its input in a list.