nixos/networkd: add `ipv4-fallback` and `fallback` as valid options to `LinkLocalAddressing`
Both options were introduced in systemd v243[1]. Those options can be
used to ensure that LinkLocalAddressing is only configured for a given
interface if DHCPv4 fails. To quote `systemd.network(5)`:

```
If "fallback" or "ipv4-fallback" is specified, then an IPv4
link-local address is configured only when DHCPv4 fails. If "fallback", an IPv6 link-local
address is always configured, and if "ipv4-fallback", the address is not configured. Note
that, the fallback mechanism works only when DHCPv4 client is enabled, that is, it requires
"DHCP=yes" or "DHCP=ipv4".
```

[1] https://github.com/systemd/systemd/commit/8bc17bb3f707c99b6c9ad295563c92b38141a1f4

(cherry picked from commit 47724fc77c7b53def3ecb66bf8c925cd17b8f069)