darling: 11b51fw -> 0.2019.8
Darling now has a release tag available on GitHub. This release includes
darlinghq/darling@c45e7204c7848729b89957f14ef1b3b742edb8af, a fix that
makes it possible to install on macOS. I'm not entirely sure why Darling
needs to be installed on macOS, but as long as it is, being able to
install it is important.

This will also switch from `fetchzip` to `fetchFromGitHub` which removes
a little of the overhead needed to install and clean up Darling.