nixos/gnome3: add gnome-flashback to systemd.packages
When we did the revert of adding gnome-flashback to systemd.packages [0]
I forgot to test with other display managers. If we use GDM with gnome-flashback
it appears it doesn't try to fallback to non-systemd startup and always fails and
starts the regular gnome-session. So adding gnome-flashback to systemd.packages
was perfectly fine, but we did forgot one detail. We need systemd targets for the
customSessions which is added using  mkSystemdTargetForWm in the gnome-
flashback package.

[0]: 42f567b30def552337bf17a6d30410c16c2b47cc