firefox-esr v52: build on Hydra
The firefox v52 is indeed unmaintained, insecure FF release, but still the
easiest way to get old Java-based sites working. My bank account site is an
example. It has vulnerabilities defined, so it is not built by Hydra.

Because of pragmatic reasons (avoid building Firefox on laptops) I propose
to bring back build on Hydra. `icedtea` project is still supported and gets
better with time, so to make use of it would be great to have a ready-to-use
browser.

Unfortunately, I can't use firefox-esr-52 from old Nixpkgs checkout
(which still has binaries in Hydra cache), as Glibc version was changed
and is now incompatible with NixOS 19.09 (segfaults).