riot-desktop: fix missing tray icon
The tray icon has always been missing for me (on sway / waybar).
It turns out that it is looked up in a different location than where we package it.

Overall, it seems we are not at all following the upstream layout as can
be seen in data.tar.xz available at packages.riot.im/debian/pool/main/r/riot-web/riot-web_1.4.2_amd64.deb

This is a small fix that works for me. May not be optimal :-).