nixos/calibre-server: Add address configuration
This commit introduces two new fields into the configuration of
calibre-server, listenAddress and port. These allow the operator to
configure the addresses that calibre should listen on, and the port it
should bind to respectively.

The default values here are copied from calibre itself, to avoid
backwards compatibility issues or more complicated service
configuration.