buildGoPackage: fix builds
Packages built with `buildGoPackage` provide a `bin` and an
`out`-output. The first one for build artifacts, the second one for
propagated build outputs. This caused a `failed to produce
output-path`-error since the new implementation to determine the
`propagaterOutput` used `dev` by default or the first listed output
(which is `bin` for go packages), hence `$out` was never created which
broke the build.

However the old implementation in `multiple-outputs.sh` didn't check for
`dev`, but for `$outputDev` which was either `dev` or `out`.

Providing `out` as fallback if `dev` doesn't exist as propagaterOutput
fixes the issue.