yosys-withGui: New wrapper derivation with GUI runtime dependencies
This commit adds a new package called 'yosys-withGui' that is
identical to yosys but wraps the bin/yosys script to setup runtime
dependencies needed for the "show" command i.e. to graphically
visualize a design.

I have consciously avoided changing the existing yosys derivation
because it takes a long time to build. The new derivation simply
copies yosys out of the Nix store and then hacks the wrapper in place.