codimd: fix build
This package broke during the update of `nodejs-10_x` to `v10.17.0`[1]
since `node-gyp` got updated from `3.8.0` to `5.0.3` during the NPM
update to `6.10.2`[2] which is bundled in the `nodejs` release tarball.

This `node-gyp`-version is used by the build script for `codimd` to
build several dependencies such as `node-sqlite3`. Older `node-gyp`
versions expected source headers for the compilation in `~/.node-gyp`,
with `node-gyp@5.0.3`, headers are declared by using the `--nodedir`
option.

[1] 7785f49b23c197ec4a02ee409b4643fa0b42b3bb
[2] https://github.com/nodejs/node/commit/e2291cf805c1dd4202367088d63e0a12a4f6e9fd