netplan: init at 0.98+git20191004
The latest stable release didn't build due to errors in their Makefile.
Going for the at this time current master branch instead. The packages
needs a bunch of cleaning up in order to be portable. CUrrently that is
done during the postPatch phase but I intend to file upstream PRs to fix
those.