pythonPackages.pony: init at 0.7.11
pony is an ORM library.

The main motivation for adding this package is to allow migrating
tribler to a newer version that supports python3.

Also adding myself to maintainers to not leave this one without, even
though I dont use it myself much, if anybody else feels like it fits their
interests, feel free to add yourselves.