mnemosyne: fix application
This application has been missing some dependencies and the
correct environment for Qt. Additionally I added a depdency
on anki to support the import feature.

The project has also moved development to Github, as very briefly
mentioned on https://mnemosyne-proj.org/contributing.php but the
2.6.1 version released there seems to differ and causes a bunch
of new problems, so leaving it for now.