ghidra: Added desktop file
Added a desktop file for Ghidra

Ghidra ships its icons as a .ico file, which isn't support by freedesktop. This means that to have icons, we need to extract the pngs from ghidra.ico, then copy them to the relevant folders.
This adds a requirement on a library to extract them, and also requires them to be copied over, which isn't ideal.