gegl_0_4: disabling jasper, fixing build
Jasper has been recently marked as insecure and removed from a bunch
of places due to some known vulnerabilities and upstream not moving
forward. This was done in https://github.com/NixOS/nixpkgs/pull/73586

This prevents gegl_0_4 from building, as jasper is a runtime dependency
per default. There is an upstream issue for that, but it does not seem
to have made progress yet: https://gitlab.gnome.org/GNOME/gegl/issues/35