pythonPackages.numpy: consistent mklSupport attribute
This commit makes `numpy` respect the `mklSupport` boolean attribute that is in
use in `pytorch` and other packages that have an MKL option.

While most packages that have this option depend on Numpy directly or
transitively, not all do; and it's more straightforward to have one config
option, since we generally want this to be applied consistently across our
library stack.