WIP: init nidaqmx
This commit adds the NI-DAQmx software for data-acquisition with NI hardware.

The nidaqmxbase package is small, however, provides functionality only for a limited set of hardware.

The nidaqmx package on the other hand supports (nearly) all hardware. Packages are provided however
only for Windows and CentOS. The RPM repo data is used to generate a package set.

TODO:
- [ ] split nidaqmx, nixdaqmxbase and nidaqmx-python in separate commits
- [ ] nidaqmx-python should optionally work with nidaqmx and/or nidaqmxbase
- [ ] fix kernel/dkms packages. Probably need to set KBUILD_EXTRA_SYMBOLS
- [ ] provide nixos module and mapping modules/packages