hello-wayland: init at 2020-07-27
This is a program that just displays a static cat picture in a Wayland
window.  I packaged it a while ago thinking it wouldn't be useful for
anybody else, but a conversation on IRC today made me realise it would
be!

hello-wayland is very useful as a minimal example when hacking on
Wayland ecosystem stuff -- even if Firefox doesn't work yet,
hello-wayland probably will and that can be useful to guide you in the
right direction!