mimalloc: un-break dynamic linking
Previous versions of the build assumed libmimalloc.so would be a hard
copy of mimalloc-secure.so iff secureBuild == true, but in 1.1.0 and
later it seems libmimalloc.so is a symlink to the -secure variant. This
apparently rectifies some behavior I noticed that was strange
previously.

This breakage wasn't caught because the 1.1.0 update was automatic in
5596317d4; there should be a checkPhase to ensure this doesn't happen
again...

Signed-off-by: Austin Seipp <aseipp@pobox.com>